requirements.txt
pika-1.1.0
pyTelegramBotAPI==3.6.7
